small crash issue fixed 'q' n'est pas reconnu en tant que commande interne ou externe, un programme excutable ou un fichier de commandes.

+++ AdminGroups/Gui/Controls/GroupItem.php
@@ -79,8 +79,14 @@ class GroupItem extends \ManiaLive\Gui\Control {
     }
 
     public function destroy() {
-        $this->plistButton->destroy();
-        $this->permiButton->destroy();
+        if($this->permiButton != null){
+            $this->permiButton->destroy();
+            $this->plistButton->destroy();
+            $this->deleteButton->destroy();
+        }
+        $this->permiButton=null;
+        $this->plistButton=null;
+        $this->deleteButton=null;
         $this->clearComponents();
